Added more error handling.

+++ src/account.py
@@ -1,6 +1,8 @@
 import random, datetime, decimal
 from src import accountdao, clientdao, bankdao
 
+ACCOUNT_TYPES = ("basic", "savings")
+
 class Account:
 	def __init__(self, accountdao, clientdao, bankdao):
 		self.client_number = clientdao.client_number
@@ -17,13 +19,21 @@ class Account:
 	@classmethod
 	def open(cls, client_number, bank_number, account_type):
 		client = clientdao.ClientDAO.readByClientNumber(client_number)
+		if client == None:
+			raise ValueError("That client doesn't exist")
 		bank = bankdao.BankDAO.readByBankNumber(bank_number)
+		if bank == None:
+			raise ValueError("That bank doesn't exist")
+		if not account_type in ACCOUNT_TYPES:
+			raise ValueError("Unknown account type")
 		account = accountdao.AccountDAO(0, client.id, bank.id, account_type, str(random.randint(100000000000, 999999999999)), False, datetime.datetime.now(), decimal.Decimal(0))
 		account = accountdao.AccountDAO.create(account)
 		return cls(account, client, bank)
 	@classmethod
 	def close(cls, account_number):
 		account = accountdao.AccountDAO.readByAccountNumber(account_number)
+		if account == None:
+			raise ValueError("That account doesn't exist")
 		accountdao.AccountDAO.delete(account)
 		return account
 	@classmethod
@@ -36,18 +46,22 @@ class Account:
 	@classmethod
 	def deposit(cls, account_number, amount):
 		if not isinstance(amount, decimal.Decimal):
-			raise TypeError()
+			raise TypeError("Amount must be a decimal")
 		if amount <= 0:
-			raise ValueError()
+			raise ValueError("Amount must be positive")
 		account = accountdao.AccountDAO.readByAccountNumber(account_number)
+		if account == None:
+			raise ValueError("That account doesn't exist")
 		account.balance += amount
 		accountdao.AccountDAO.update(account)
 	@classmethod
 	def withdraw(cls, account_number, amount):
 		if not isinstance(amount, decimal.Decimal):
-			raise TypeError()
+			raise TypeError("Amount must be a decimal")
 		if amount <= 0:
-			raise ValueError()
+			raise ValueError("Amount must be positive")
 		account = accountdao.AccountDAO.readByAccountNumber(account_number)
+		if account == None:
+			raise ValueError("That account doesn't exist")
 		account.balance -= amount
 		accountdao.AccountDAO.update(account)
